Custom mainmenu.nib doesn't expose cmd-W
The command key is a default keybinding in the game on macOS, and the W key may be used as an alternative to the up arrow.

However, hitting cmd+W will trigger "Close Window", causing the game to quit.

SDL_cocoaevents.m tries to load the .nib file referenced in info.plist (key "NSMainNibFile"). If this file is missing, SDL will build a hardcoded "Window" menu containing a "Close Window" entry.

So, we're supplying our own nib file without a cmd+W shortcut. This way, the shortcut won't interfere with normal gameplay on QWERTY keyboards.

We're keeping cmd+Q as this is an expected shortcut in all Mac applications. This may interfere with AZERTY keyboards, though (where Q can be interpreted as the 'left' key in a ZQSD cluster).